scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "StylusPort Scripts"
authors = [
    { name = "Chris Ricketts", email = "chris_ricketts@protonmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
